refactoring product
